+++ packages/react/src/server.ts
@@ -1,10 +1,19 @@
+import {
+  defineRender,
+  getComponentDescriptor,
+  type ComponentProps,
+} from "@web-widget/schema/server-helpers";
+import type { ReactNode } from "react";
+import { Fragment, createElement } from "react";
 import { __ENV__ } from "./web-widget";
-import { Fragment } from "react";
-import { type Attributes, createElement } from "react";
+
+import type {
+  ReactDOMServerReadableStream,
+  RenderToReadableStreamOptions,
+} from "react-dom/server";
 // @ts-ignore
 import * as ReactDOMServer from "react-dom/server.browser";
-import { defineRender } from "@web-widget/schema/server-helpers";
-import type { DefineVueRenderOptions } from "./types";
+import type { DefineReactRenderOptions } from "./types";
 
 export * from "@web-widget/schema/server-helpers";
 export * from "./web-widget";
@@ -28,43 +37,67 @@ function htmlEscapeJsonString(str: string): string {
   return str.replace(ESCAPE_REGEX, (match) => ESCAPE_LOOKUP[match]);
 }
 
+function renderToReadableStream(
+  vNode: ReactNode,
+  renderOptions?: RenderToReadableStreamOptions
+): Promise<ReactDOMServerReadableStream> {
+  return ReactDOMServer.renderToReadableStream(vNode, renderOptions);
+}
+
+export interface ReactRenderOptions {
+  react?: {
+    awaitAllReady?: boolean;
+  } & RenderToReadableStreamOptions;
+}
+
 export const defineReactRender = ({
   onPrefetchData,
-}: DefineVueRenderOptions = {}) => {
-  return defineRender(async (context, component, props) => {
-    const state = onPrefetchData
-      ? await onPrefetchData(context, component, props)
-      : undefined;
-    const mergedProps = state ? Object.assign({}, state, props) : props;
+}: DefineReactRenderOptions = {}) => {
+  return defineRender<ReactRenderOptions>(
+    async (context, { react: reactRenderOptions } = {}) => {
+      const componentDescriptor = getComponentDescriptor(context);
+      const { component, props } = componentDescriptor;
+      const state = onPrefetchData
+        ? await onPrefetchData(context, component, props)
+        : undefined;
+      const mergedProps = (
+        state ? Object.assign({}, state, props) : props
+      ) as ComponentProps<any>;
 
-    let vNode;
-    if (
-      typeof component === "function" &&
-      component.constructor.name === "AsyncFunction"
-    ) {
-      // experimental
-      vNode = await component(mergedProps);
-    } else {
-      vNode = createElement(component, mergedProps as Attributes);
-    }
+      let vNode;
+      if (
+        typeof component === "function" &&
+        component.constructor.name === "AsyncFunction"
+      ) {
+        // experimental
+        vNode = await component(mergedProps);
+      } else {
+        vNode = createElement(component, mergedProps);
+      }
 
-    if (state) {
-      return ReactDOMServer.renderToReadableStream(
-        createElement(Fragment, null, [
-          vNode,
-          createElement("script", {
-            as: "state",
-            type: "application/json",
-            dangerouslySetInnerHTML: {
-              __html: htmlEscapeJsonString(JSON.stringify(state)),
-            },
-          }),
-        ])
-      );
-    } else {
-      return ReactDOMServer.renderToReadableStream(vNode);
+      const stream = state
+        ? await renderToReadableStream(
+            createElement(Fragment, null, [
+              vNode,
+              createElement("script", {
+                as: "state",
+                type: "application/json",
+                dangerouslySetInnerHTML: {
+                  __html: htmlEscapeJsonString(JSON.stringify(state)),
+                },
+              }),
+            ]),
+            reactRenderOptions
+          )
+        : await renderToReadableStream(vNode, reactRenderOptions);
+
+      if (reactRenderOptions?.awaitAllReady) {
+        await stream.allReady;
+      }
+
+      return stream;
     }
-  });
+  );
 };
 
 export const render = defineReactRender();